{{cookiecutter.__repo_name}}/requirements.txt
boto3
botocore
imbalanced-learn
google-cloud-storage
google-cloud-bigquery
matplotlib
mlflow
numpy
openpyxl
pandas
python-dotenv
pyyaml
requests
scikit-learn
seaborn
snowflake-connector-python[pandas]
tqdm
xgboost
xlsxwriter